+++ client/httplib/module.go
@@ -1,4 +1,4 @@
-// Copyright 2020 beego
+// Copyright 2021 beego
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,4 +14,4 @@
 
 package httplib
 
-
+const moduleName = "httplib"